Revert "Add support or .NET 4.5 (#9)"
This reverts commit 22df766c92c7659077362774116a4df3a21f6172.
