+++ .vscode/settings.json
@@ -1,0 +1,16 @@
+// Place your settings in this file to overwrite the default settings
+{ 
+    "window.openFilesInNewWindow": "on",
+    "git.enabled": true,
+    "files.exclude": {
+        "**/.git": true,
+        "**/.DS_Store": true,
+        "**/npm-debug.log": true,
+        "**/bin": true,
+        "**/obj": true,
+    },
+    "git.confirmSync": false,
+    "workbench.colorTheme": "Monokai",
+    "editor.renderWhitespace": "all",
+    "git.autofetch": true
+}
