+++ HL7-dotnetcore.sln
@@ -1,10 +1,11 @@
+
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26124.0
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "hl7-dotnetcore", "src\hl7-dotnetcore.csproj", "{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "hl7-dotnetcore", "src\hl7-dotnetcore.csproj", "{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "hl7-tests", "tests\hl7-tests\hl7-tests.csproj", "{A9E9C013-5954-4CCA-8917-AFE7CEB47A4A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "hl7test", "test\hl7test.csproj", "{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,36 +16,33 @@ Global
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|x64.Build.0 = Debug|Any CPU
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|x86.Build.0 = Debug|Any CPU
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|x64.ActiveCfg = Release|Any CPU
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|x64.Build.0 = Release|Any CPU
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|x86.ActiveCfg = Release|Any CPU
-		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|x86.Build.0 = Release|Any CPU
-		{A9E9C013-5954-4CCA-8917-AFE7CEB47A4A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A9E9C013-5954-4CCA-8917-AFE7CEB47A4A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A9E9C013-5954-4CCA-8917-AFE7CEB47A4A}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{A9E9C013-5954-4CCA-8917-AFE7CEB47A4A}.Debug|x64.Build.0 = Debug|Any CPU
-		{A9E9C013-5954-4CCA-8917-AFE7CEB47A4A}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{A9E9C013-5954-4CCA-8917-AFE7CEB47A4A}.Debug|x86.Build.0 = Debug|Any CPU
-		{A9E9C013-5954-4CCA-8917-AFE7CEB47A4A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A9E9C013-5954-4CCA-8917-AFE7CEB47A4A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A9E9C013-5954-4CCA-8917-AFE7CEB47A4A}.Release|x64.ActiveCfg = Release|Any CPU
-		{A9E9C013-5954-4CCA-8917-AFE7CEB47A4A}.Release|x64.Build.0 = Release|Any CPU
-		{A9E9C013-5954-4CCA-8917-AFE7CEB47A4A}.Release|x86.ActiveCfg = Release|Any CPU
-		{A9E9C013-5954-4CCA-8917-AFE7CEB47A4A}.Release|x86.Build.0 = Release|Any CPU
-	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {DDA1F5BF-CD42-47C7-B807-43638D788A1B}
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|x64.ActiveCfg = Debug|x64
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|x64.Build.0 = Debug|x64
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|x86.ActiveCfg = Debug|x86
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Debug|x86.Build.0 = Debug|x86
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|x64.ActiveCfg = Release|x64
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|x64.Build.0 = Release|x64
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|x86.ActiveCfg = Release|x86
+		{8D0228F0-7BE1-4D45-9EB4-8E170F76A17E}.Release|x86.Build.0 = Release|x86
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Debug|x64.ActiveCfg = Debug|x64
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Debug|x64.Build.0 = Debug|x64
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Debug|x86.ActiveCfg = Debug|x86
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Debug|x86.Build.0 = Debug|x86
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Release|x64.ActiveCfg = Release|x64
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Release|x64.Build.0 = Release|x64
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Release|x86.ActiveCfg = Release|x86
+		{910F6F49-6AD7-45EF-8FA0-E141C77D81DA}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 EndGlobal
