+++ src/HL7-dotnetcore.csproj
@@ -3,7 +3,7 @@
     <Description>Small HL7 2.x library compatible with .net Core</Description>
     <Copyright>(c) Efferent Health, LLC</Copyright>
     <VersionPrefix>2.5.0</VersionPrefix>
-    <TargetFrameworks>netstandard1.3;netstandard1.5;netstandard1.6;netstandard2.0;net45</TargetFrameworks>
+    <TargetFrameworks>netstandard1.3;netstandard1.5;netstandard1.6;netstandard2.0</TargetFrameworks>
     <AssemblyName>HL7-dotnetcore</AssemblyName>
     <PackageId>HL7-dotnetcore</PackageId>
     <PackageTags>hl7;dotnetcore</PackageTags>
